Reorg of main stepper to allow different steppers. Split out the present wave_prop/single-step method to its own file. Added a dummy for SSP-RK3.
